feat/#36 : types 분리
