+++ apis/exp.ts
@@ -1,23 +1,11 @@
 'use server';
 
+import {
+  ExperienceFormType,
+  ExperienceStatus,
+  ExperienceType,
+} from '@/types/exp';
 import API from './config';
-
-export type ExperienceStatus = 'DRAFT' | 'SAVE';
-
-export type ExperienceType =
-  | 'CONTEST'
-  | 'EXTERNAL_ACTIVITIES'
-  | 'ACADEMIC_CLUB'
-  | 'INTERN'
-  | 'PROJECT'
-  | 'EDUCATION'
-  | 'PRIZE'
-  | 'CERTIFICATES'
-  | 'VOLUNTEER_WORK'
-  | 'STUDY_ABROAD'
-  | 'ETC';
-
-export type ExperienceFormType = 'STAR_FORM' | 'SIMPLE_FORM';
 
 export interface ExperiencePayload {
   status: ExperienceStatus;
