+++ types/exp.ts
@@ -1,0 +1,16 @@
+export type ExperienceStatus = 'DRAFT' | 'SAVE';
+
+export type ExperienceType =
+  | 'CONTEST'
+  | 'EXTERNAL_ACTIVITIES'
+  | 'ACADEMIC_CLUB'
+  | 'INTERN'
+  | 'PROJECT'
+  | 'EDUCATION'
+  | 'PRIZE'
+  | 'CERTIFICATES'
+  | 'VOLUNTEER_WORK'
+  | 'STUDY_ABROAD'
+  | 'ETC';
+
+export type ExperienceFormType = 'STAR_FORM' | 'SIMPLE_FORM';
